refactor: align project structure with vercel conventions
