+++ src/app.ts
@@ -1,9 +1,9 @@
 import express from 'express';
 import 'express-async-errors'; 
-//import path from 'path'; 
+import path from 'path'; 
 import { setupMiddleware } from './middleware';
-//import routes from './routes/index';
-//import { errorHandlerMiddleware } from './middlewares/error-handler';
+import routes from './routes/index';
+import { errorHandlerMiddleware } from './middlewares/error-handler';
 
 const app = express();
 
@@ -19,11 +19,10 @@ app.get('/health', (req, res) => {
 // This allows the frontend to access saved images via URL
 // http://localhost:8080/uploads/nome-da-imagem.jpg
 
+app.use('/uploads', express.static(path.join('/tmp', 'uploads')));
 
-//app.use('/uploads', express.static(path.join('/tmp', 'uploads')));
+app.use(routes);
 
-//app.use(routes);
-
-//app.use(errorHandlerMiddleware);
+app.use(errorHandlerMiddleware);
 
 export default app;
